+++ internal/cgo/cgo_disabled.go
@@ -1,6 +1,0 @@
-//go:build !cgo
-
-package cgo
-
-// Enabled is set to true if CGO is enabled.
-var Enabled = false