+++ internal/compilation/cgo_disabled.go
@@ -1,0 +1,6 @@
+//go:build !cgo
+
+package compilation
+
+// CgoEnabled is set to false if CGO is disabled.
+var CgoEnabled = false