+++ internal/compilation/cgo_enabled.go
@@ -1,0 +1,6 @@
+//go:build cgo
+
+package compilation
+
+// CgoEnabled is set to true if CGO is enabled.
+var CgoEnabled = true